Room construction
Arbitrary way of populating room coordinates// starting and ending door locations for now

Need to work on amt of rooms- logic for linking rooms together upon instantiation. If we populate (construct) all of the doors at once, this would create an issue with needing to create another dimension.. A good way to do that will be to have it so once a door is closed or a room is left- and that room is decomposed and an entirely new room is created. The logic behind can be tied together another way unless someone has another suggestion!